+++ agent/brookesia_agent_coze/src/agent_coze.cpp
@@ -176,6 +176,8 @@ bool Coze::set_info(const boost::json::object &info)
 
     BROOKESIA_LOGD("Params: info(%1%)", BROOKESIA_DESCRIBE_TO_STR(info));
 
+    try_load_data();
+
     CozeInfo coze_info;
 
     auto success = BROOKESIA_DESCRIBE_FROM_JSON(info, coze_info);
@@ -188,7 +190,7 @@ bool Coze::set_info(const boost::json::object &info)
     auto current_info_str = BROOKESIA_DESCRIBE_JSON_SERIALIZE(get_data<DataType::Info>());
     auto new_info_str = BROOKESIA_DESCRIBE_JSON_SERIALIZE(coze_info);
     if (current_info_str == new_info_str) {
-        BROOKESIA_LOGD("Info is the same, skip");
+        BROOKESIA_LOGI("Info is the same, skip setting");
         return true;
     }