+++ agent/brookesia_agent_openai/include/brookesia/agent_openai/macro_configs.h
@@ -8,7 +8,7 @@
 #include "sdkconfig.h"
 
 /* Debug log */
-#define BROOKESIA_AGENT_OPENAI_LOG_TAG "AgentOpenai"
+#define BROOKESIA_AGENT_OPENAI_LOG_TAG "Openai"
 
 #if !defined(BROOKESIA_AGENT_OPENAI_ENABLE_DEBUG_LOG)
 #   if defined(CONFIG_BROOKESIA_AGENT_OPENAI_ENABLE_DEBUG_LOG)